pyproject.toml
[tool.poetry]
name = "threes"
version = "0.0.1"
description = "Instagram 3x1 spacer"
authors = ["Andrew Wong <andrew.j.wong@outlook.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
instagram-private-api = {git = "https://git@github.com/ping/instagram_private_api.git", rev = "1.6.0"}
python-dotenv = "^0.17.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
